requirements.txt
blinker==1.6.2
Brotli==1.1.0
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Compress==1.13
Flask-Mail==0.9.1
Flask-Scss==0.5
Flask-WTF==1.1.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pyScss==1.4.0
python-dotenv==1.0.0
six==1.16.0
Werkzeug==2.3.6
WTForms==3.0.1
zipp==3.15.0
